requirements.txt
pyspark==3.5.0
azure-identity==1.15.0
azure-core==1.29.6
azure-storage-blob==12.19.0
